adapted createTask-steps, moved all to new PO hometaskPages and deleted old createtask.js.
