+++ features/homework.feature
@@ -1,4 +1,5 @@
-@createHomework
+@
+
 Feature: create different types of homework
