+++ page-objects/pages/HMWRKAddEditHomeworkPage.js
@@ -3,51 +3,95 @@
 const waitHelpers = require('../../runtime/helpers/waitHelpers.js');
 const dateTimeHelpers = require('../../runtime/helpers/dateTimeHelpers.js');
 const courseData = require('../../shared-objects/courseData');
-const cRSSCourseHomeworksPage = require("../pages/coursePages/CRSSCourseHomeworksPage.js");
+const courseHomeworksPage = require("../pages/coursePages/CRSSCourseHomeworksPage.js");
 
-module.exports = {
-    setPrivate: async function () {
-        await waitHelpers.waitAndClick(courseData.elem.checkbox);
-    },
-    setHometaskText: async function () {
-        await driver.pause(global.SHORT_WAIT_MILLIS);
-        await driver.switchToFrame(0);
-        let body = await driver.$('body');
-        let message = 'Here is some TEXT!';
-        await body.setValue(message);
-        await driver.switchToParentFrame();
-    },
-    setAccomplishTime: async function () {
-        var begin = await dateTimeHelpers.dateToString();
-        await driver.execute(`document.querySelector("#availableDate").value="${begin}"`);
-        var end = await dateTimeHelpers.randomDate();
-        await driver.execute(`document.querySelector("#dueDate").value="${end}"`);
-    },
-    clickAdd: async function () {
-        let container = await driver.$('#homework-form');
-        let addBtn = await container.$('button[type="submit"]');
-        await addBtn.click();
-        let selectorToBeLoaded = await driver.$('#homeworks');
-        await selectorToBeLoaded.waitForExist(2000);
-    },
+const taskNameField = "input[placeholder='Titel']";
+const submitTaskButton = ".btn-submit";
+
+    module.exports = {
+
+        setPrivate: async function () {
+
+            await waitHelpers.waitAndClick(courseData.elem.checkbox);
+
+        },
+
+        setHomeworkName: async function(taskName) {
+            const nameField = await driver.$(taskNameField);
+            await nameField.setValue(taskName);
+        },
+
+        setHometaskText: async function () {
+
+            await driver.pause(global.SHORT_WAIT_MILLIS);
+            await driver.switchToFrame(0);
+            let body = await driver.$('body');
+            let message = 'Here is some TEXT!';
+            await body.setValue(message);
+            await driver.switchToParentFrame();
+
+        },
+        setAccomplishTime: async function () {
+
+            var begin = await dateTimeHelpers.dateToString();
+            await driver.execute(`document.querySelector("#availableDate").value="${begin}"`);
+            var end = await dateTimeHelpers.randomDate();
+            await driver.execute(`document.querySelector("#dueDate").value="${end}"`);
+
+        },
+        clickAdd: async function () {
+
+            let container = await driver.$('#homework-form');
+            let addBtn = await container.$('button[type="submit"]');
+            await addBtn.click();
+            let selectorToBeLoaded = await driver.$('#homeworks');
+            await selectorToBeLoaded.waitForExist(2000);
+
+        },
+
+        clickSubmitTaskButton: async function() {
+            await waitHelpers.waitAndClick(submitTaskButton);
+        },
+
+        addBasicHometask: async function (coursename, taskname) {
+
+            await courseHomeworksPage.clickCreateNewTaskInTheCourse(coursename);
+            let nameSelector = await driver.$(courseData.elem.homeworkName);
+            await nameSelector.setValue(taskname);
+            await waitHelpers.waitAndClick(courseData.elem.teamworkActivate);
+            await this.setAccomplishTime();
+            await this.setHometaskText();
+            await waitHelpers.waitAndClick(courseData.elem.submitAddHomeworkBtn);
+
+        },
+        addPrivateHometask: async function (coursename, taskname) {
+
+            await courseHomeworksPage.clickCreateNewTaskInTheCourse(coursename);
+            let nameSelector = await driver.$(courseData.elem.homeworkName);
+            await nameSelector.setValue(taskname);
+            await waitHelpers.waitAndClick(courseData.elem.teamworkActivate);
+            await this.setAccomplishTime();
+            await this.setHometaskText();
+            await this.setPrivate();
+            await waitHelpers.waitAndClick(courseData.elem.submitAddHomeworkBtn);
+
+        },
     
-	addBasicHometask: async function (coursename, taskname) {
-		await cRSSCourseHomeworksPage.clickCreateNewTaskInTheCourse(coursename);
-		let nameSelector = await driver.$(courseData.elem.homeworkName);
-		await nameSelector.setValue(taskname);
-		await waitHelpers.waitAndClick(courseData.elem.teamworkActivate);
-		await this.setAccomplishTime();
-		await this.setHometaskText();
-		await waitHelpers.waitAndClick(courseData.elem.submitAddHomeworkBtn);
-	},
-	addPrivateHometask: async function (coursename, taskname) {
-		await cRSSCourseHomeworksPage.clickCreateNewTaskInTheCourse(coursename);
-		let nameSelector = await driver.$(courseData.elem.homeworkName);
-		await nameSelector.setValue(taskname);
-		await waitHelpers.waitAndClick(courseData.elem.teamworkActivate);
-		await this.setAccomplishTime();
-		await this.setHometaskText();
-		await this.setPrivate();
-		await waitHelpers.waitAndClick(courseData.elem.submitAddHomeworkBtn);
-	},
-}
+        getTaskNames: async function() {
+            await driver.pause(1000 * 10);
+            const container = await driver.$(".col-xl-12");
+            const tasksArray = await container.$$("li");
+            const namesArray = [];
+            for (var i = 1; i <= tasksArray.length; i++) {
+                const task = await container.$(
+                    "li:nth-child(" + i + ") > a:nth-child(1)"
+                );
+                const taskName = (await task.getAttribute("title"))
+                    .replace("Details der Aufgabe: '", "")
+                    .replace(/'$/gi, "");
+                await namesArray.push(taskName);
+            }
+            return namesArray;
+        }
+
+    }
