+++ step_definitions/createTask-steps.js
@@ -1,12 +1,15 @@
 "use strict";
 
-const createTask = require("../page-objects/createTask");
+
 const loginData = require("../shared-objects/loginData");
 const firstLogin = require("../shared_steps/firstLogin.js");
 const loginPage = require('../page-objects/pages/generalPagesBeforeLogin/LoginPage.js');
 const startPage = require('../page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js');
 const elementHelpers = require('../runtime/helpers/elementHelpers.js');
-let page = { createTask };
+const navigationLeftPage = require('../page-objects/pages/NavigationLeftPage.js');
+const homeworkListPage = require('../page-objects/pages/HMWRKHomeworkListPage.js');
+const addEditHomeworkPage = require('../page-objects/pages/HMWRKAddEditHomeworkPage.js');
+
 
 Given(/^teacher goes to the home page$/, function() {
 	return elementHelpers.loadPage(loginData.url, 20);
@@ -23,38 +26,38 @@ When(/^the teacher has accepted the data protection agreement$/, function() {
 });
 
 Then(/^he should click the task-button in the dashboard-sidebar$/, function() {
-	return page.createTask.clickSidebarTaskButton();
+	return navigationLeftPage.clickNavItemTasks();
 });
 
 Then(/^he should click the create-task-button on the task page$/, function() {
-	return page.createTask.clickCreateTaskButton();
+	return homeworkListPage.clickCreateTaskButton();
 });
 
 Then(/^he should put the taskname (.*) into the name field$/, function(
 	taskName
 ) {
-	return page.createTask.putTaskName(taskName);
+	return addEditHomeworkPage.setHomeworkName(taskName);
 });
 
 Then(/^he should put the taskBody (.*) into the body field$/, function(
 	taskBody
 ) {
-	return page.createTask.putTaskBody(taskBody);
+	return addEditHomeworkPage.setHometaskText(taskBody);
 });
 
 Then(
 	/^he should click the submit-task-button on the task-creation-form$/,
 	function() {
-		return page.createTask.clickSubmitTaskButton();
+		return addEditHomeworkPage.clickSubmitTaskButton();
 	}
 );
 Then(/^teacher goes to tasks page$/, function() {
-	return page.createTask.goToTasks();
+	return navigationLeftPage.clickNavItemTasks();
 x});
 
 Then(
 	/^he should see the created task with the name (.*) on the task page$/,
 	async function(taskName) {
-		let tasks = await page.createTask.getTaskNames();
+		let tasks = await addEditHomeworkPage.getTaskNames();
 		await expect(taskName).to.be.oneOf(tasks);
 	});
