+++ step_definitions/navigationLeft-steps.js
@@ -4,8 +4,10 @@ const leftNavigation = require('../page-objects/pages/NavigationLeftPage.js');
 const leftMenuItems = leftNavigation.navItemsEnum;
 
 Then('click left navigation item {string}', function (string) {
+
     let navigationItem = string;
     switch (navigationItem) {
+
         case leftMenuItems.LOGO:
             return leftNavigation.clickNavItemLogo();
         case leftMenuItems.DASHBOARD:
@@ -66,5 +68,7 @@ Then('click left navigation item {string}', function (string) {
             return leftNavigation.clickNavItemHelpContact();
         default:
             return Error("no such element found in 'click left navigation item {string}' " + navigationItem);
+
     }
+
 });
